requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
Flask>=1,<2
python-jose-cryptodome
flask-cors
requests
azure-identity
azure-functions
azure-mgmt-compute
azure-mgmt-resource
azure-mgmt-network